// src/gen/bikehub/model/createEquipment.ts
/**
 * BikeHubParthenope
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EquipmentType } from './equipmentType';


export interface CreateEquipment { 
    bicycleId?: string;
    type: EquipmentType;
    name: string;
    description: string;
    imageUrl?: string;
}