// src/gen/bikehub/model/equipment.ts
/**
 * BikeHubParthenope
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EquipmentType } from './equipmentType';


export interface Equipment { 
    id: string;
    bicycleId?: string;
    type: EquipmentType;
    name: string;
    description: string;
    imageUrl?: string;
}