// src/gen/bikehub/model/equipmentsPage.ts
/**
 * BikeHubParthenope
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Equipment } from './equipment';


export interface EquipmentsPage { 
    totalItems?: number;
    currentPage?: number;
    totalPages?: number;
    results?: Array<Equipment>;
}